Code for a generic Partial Information Decomposition.
Two methods are currently implemented: CVX, which needs the CVX package, and linprog. The first is more reliable, due to the built-in CVX features, and could be optimised top-down (tweaking CVX parameters or code to get what we want); the second is faster but less reliable, and could be optimised bottom-up (adding more features to the standard linprog MATLAB function).